requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

